cli: use a more idiomatic and simple package layout
* We renamed the `pkg` package to `internal`. `pkg` is not a standard
  naming and doesn't mean a lot. Nowadays, usually people put things
  under `internal`, because it indicates that only internal logic
  is placed there.
* Moved all root level packages under internal. In example, `auth`,
  `config`, `printer`, `testutil`, etc.. are all now under the
  `internal` folder

The new layout produces the following layout:

```
.
├── cmd
│   └── pscale/main.go
├── internal
│   ├── auth/u
│   ├── cmd
│   ├── cmdutil
│   ├── config
│   ├── printer
│   ├── promptutil
│   ├── proxyutil
│   └── testutil

```

There are still thing we can do, such as making sure `printer`,
`cmdutil` and `promptutil` are merged. I feel like tthey all belong into
a `cmdutil` package, with utility functions for the commands.
