+++ internal/printer/schema_snapshot.go
