+++ internal/promptutil/branch.go
