--- pkg/cmd/org/org.go
@@ -1,16 +1,0 @@
-package org
-
-import (
-	"github.com/planetscale/cli/config"
-	"github.com/spf13/cobra"
-)
-
-func OrgCmd(cfg *config.Config) *cobra.Command {
-	cmd := &cobra.Command{
-		Use: "org <command>",
-	}
-
-	cmd.AddCommand(SwitchCmd(cfg))
-
-	return cmd
-}
